+++ examples/test.h
@@ -1,0 +1,6 @@
+#ifndef TEST_HDR_H
+#define TEST_HDR_H
+
+#include "lwbtn/lwbtn.h"
+
+#endif /* TEST_HDR_H */