+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Develop
 
+## v1.0.0
+
+- Send `CLICK` event if there is an overlap between max time between clicks and new click arrives
+- Do not send `CLICK` event if there was previously detected long hold press (hold time exceeded max allowed click time)
+
 ## v0.0.2
 
 - Add `LWBTN_CFG_GET_STATE_MODE` to control *get state* mode